requirements.txt
streamlit==1.40.0
requests==2.32.3
python-dotenv==1.0.1
certifi==2024.2.2
